# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Quinn AI Personal Assistant API
  description: |
    Quinn is an AI-powered personal assistant that helps manage emails, banking, 
    food planning, life tasks, and more. All endpoints require authentication via 
    AWS Cognito JWT tokens.
  version: 1.0.0
  contact:
    name: Quinn Support
    email: support@quinn.ai
servers:
  - url: https://8aqsagpkp6.execute-api.eu-west-2.amazonaws.com/dev
    description: Development server
  - url: https://api.quinn.ai
    description: Production server (future)
security:
  - bearerAuth: []
...
paths:
...
  /calendar/events:
    get:
      tags: [Calendar]
      summary: Get calendar events
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
        - name: maxResults
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Events retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CalendarEvent"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Calendar not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Calendar]
      summary: Create calendar event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCalendarEventRequest"
      responses:
        "201":
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CalendarEvent"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Calendar not connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"

  # Calls
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: AWS Cognito JWT token
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Validation error"
            details: "Field 'title' is required"
    Unauthorized:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Unauthorized"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            message: "Internal server error"
  schemas:
    Error:
      type: object
      required:
        - message
      properties:
        message:
          type: string
        details:
          type: string
        code:
          type: string
...
    CalendarEvent:
      type: object
      properties:
        eventId:
          type: string
        title:
          type: string
        description:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        location:
          type: string
        attendees:
          type: array
          items:
            type: string
        isAllDay:
          type: boolean
    CreateCalendarEventRequest:
      type: object
      required:
        - title
        - start
        - end
      properties:
        title:
          type: string
        description:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: string
          format: date-time
        location:
          type: string
        attendees:
          type: array
          items:
            type: string
...